refactor(playlist-maker): rename some files
