+++ custom-apps/playlist-maker/src/models/nodes/filter/tempo-processor.ts
@@ -1,6 +1,6 @@
 import { setAudioFeatures } from 'custom-apps/playlist-maker/src/utils/track-utils';
 import { z } from 'zod';
-import { type WorkflowTrack } from '../../track';
+import { type WorkflowTrack } from '../../workflow-track';
 import { BaseNodeDataSchema, NodeProcessor } from '../node-processor';
 
 export const MIN_TEMPO = 0;
