+++ custom-apps/playlist-maker/src/models/nodes/processing/shuffle-processor.ts
@@ -1,4 +1,4 @@
-import { type WorkflowTrack } from '../../track';
+import { type WorkflowTrack } from '../../workflow-track';
 import { type BaseNodeData, NodeProcessor } from '../node-processor';
 
 export class ShuffleProcessor extends NodeProcessor<BaseNodeData> {
