+++ custom-apps/playlist-maker/src/models/nodes/sources/album-source-processor.ts
@@ -2,7 +2,7 @@ import type { SimpleTrack } from '@shared/components/track-list/models/interface
 import { GRAPHQL_MAX_LIMIT } from '@shared/graphQL/constants';
 import { getAlbum, type GetAlbumData } from '@shared/graphQL/queries/get-album';
 import { z } from 'zod';
-import { type WorkflowTrack } from '../../track';
+import { type WorkflowTrack } from '../../workflow-track';
 import { BaseNodeDataSchema, NodeProcessor } from '../node-processor';
 
 export const AlbumDataSchema = z
