+++ custom-apps/playlist-maker/src/models/nodes/sources/playlist-tracks-source-processor.ts
@@ -5,7 +5,7 @@ import {
 } from '@shared/platform/playlist';
 import { getPlatform } from '@shared/utils/spicetify-utils';
 import { z } from 'zod';
-import { type WorkflowTrack } from '../../track';
+import { type WorkflowTrack } from '../../workflow-track';
 import { BaseNodeDataSchema, NodeProcessor } from '../node-processor';
 
 export const PlaylistDataSchema = z
