+++ custom-apps/playlist-maker/src/models/nodes/sources/radio-source-processor.ts
@@ -6,7 +6,7 @@ import {
 import { getRadioPlaylist } from '@shared/spclient/get-radio-playlist';
 import { getPlatform } from '@shared/utils/spicetify-utils';
 import { z } from 'zod';
-import { type WorkflowTrack } from '../../track';
+import { type WorkflowTrack } from '../../workflow-track';
 import { BaseNodeDataSchema, NodeProcessor } from '../node-processor';
 
 export const RadioDataSchema = z
