+++ custom-apps/playlist-maker/src/models/nodes/sources/top-tracks-source-processor.ts
@@ -5,7 +5,7 @@ import {
 import { type Track as ApiTrack } from '@shared/api/models/track';
 import { getAllPages } from '@shared/utils/web-api-utils';
 import { z } from 'zod';
-import { type WorkflowTrack } from '../../track';
+import { type WorkflowTrack } from '../../workflow-track';
 import { BaseNodeDataSchema, NodeProcessor } from '../node-processor';
 
 export const TopTracksDataSchema = z
