+++ custom-apps/playlist-maker/src/stores/store.ts
@@ -17,7 +17,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { create } from 'zustand';
 import { type BaseNodeData } from '../models/nodes/node-processor';
 import { type CustomNodeType } from '../models/nodes/node-types';
-import { type WorkflowTrack } from '../models/track';
+import { type WorkflowTrack } from '../models/workflow-track';
 import { getDefaultValueForNodeType } from '../utils/node-utils';
 import type { SavedWorkflow } from '../utils/storage-utils';
